refactor: prod 환경은 Polling 메서드 로깅하지 않도록 수정 #251
